+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [api] Only allow specific routes in epehemeral mode
+- [ui] Fix index input in Elasticsearch pipeline output form
 - [ui] Login form when only MRT is available
 - [ui] Always redirect to home page in epehemeral mode